move flr serialization tests to kernel tests
