+++ tensorflow/python/data/kernel_tests/fixed_length_record_dataset_test.py
@@ -24,6 +24,8 @@ import zlib
 
 from absl.testing import parameterized
 
+from tensorflow.python.data.experimental.kernel_tests import reader_dataset_ops_test_base
+from tensorflow.python.data.kernel_tests import checkpoint_test_base
 from tensorflow.python.data.kernel_tests import test_base
 from tensorflow.python.data.ops import readers
 from tensorflow.python.framework import combinations
@@ -208,6 +210,23 @@ class FixedLengthRecordDatasetTest(test_base.DatasetTestBase,
           [self._record(j, i) for i in range(self._num_records)])
     self.assertDatasetProduces(dataset, expected_output=expected_output,
                                assert_items_equal=True)
+
+class FixedLengthRecordDatasetCheckpointTest(
+    reader_dataset_ops_test_base.FixedLengthRecordDatasetTestBase,
+    checkpoint_test_base.CheckpointTestBase, parameterized.TestCase):
+
+  def _build_iterator_graph(self, num_epochs, compression_type=None):
+    filenames = self._createFiles()
+    return readers.FixedLengthRecordDataset(
+        filenames, self._record_bytes, self._header_bytes,
+        self._footer_bytes).repeat(num_epochs)
+
+  @combinations.generate(test_base.default_test_combinations())
+  def testFixedLengthRecordCore(self):
+    num_epochs = 5
+    num_outputs = num_epochs * self._num_files * self._num_records
+    self.run_core_tests(lambda: self._build_iterator_graph(num_epochs),
+                        num_outputs)
 
 
 if __name__ == "__main__":
